设置 createuser

+++ src/main/java/com/ldps/data/ResourceData.java
@@ -21,7 +21,7 @@ public class ResourceData {
     
     private String createDateStr;
     
-    private Integer createUser;
+    private Long createUser;
 
     private Integer buildingId;
     
@@ -204,11 +204,11 @@ public class ResourceData {
         this.createDate = createDate;
     }
 
-    public Integer getCreateUser() {
+    public Long getCreateUser() {
         return createUser;
     }
 
-    public void setCreateUser(Integer createUser) {
+    public void setCreateUser(Long createUser) {
         this.createUser = createUser;
     }
