+++ src/main/java/com/ldps/data/ResourceGroupData.java
@@ -14,7 +14,7 @@ public class ResourceGroupData {
 
     private Date createDate;
 
-    private Integer createUser;
+    private Long createUser;
 
     private Long num1;
 
@@ -68,11 +68,11 @@ public class ResourceGroupData {
         this.createDate = createDate;
     }
 
-    public Integer getCreateUser() {
+    public Long getCreateUser() {
         return createUser;
     }
 
-    public void setCreateUser(Integer createUser) {
+    public void setCreateUser(Long createUser) {
         this.createUser = createUser;
     }
