+++ src/main/java/com/ldps/model/ResourceGroupModel.java
@@ -14,7 +14,7 @@ public class ResourceGroupModel {
 
     private Date createDate;
 
-    private Integer createUser;
+    private Long createUser;
 
     private Long num1;
 
@@ -109,11 +109,11 @@ public class ResourceGroupModel {
         this.createDate = createDate;
     }
 
-    public Integer getCreateUser() {
+    public Long getCreateUser() {
         return createUser;
     }
 
-    public void setCreateUser(Integer createUser) {
+    public void setCreateUser(Long createUser) {
         this.createUser = createUser;
     }
